fix: removes the retry behaviour from rs.reconfig
