+++ packages/shell-api/src/replica-set.spec.ts
@@ -4,7 +4,7 @@ import chai, { expect } from 'chai';
 import { EventEmitter } from 'events';
 import semver from 'semver';
 import sinonChai from 'sinon-chai';
-import sinon, { StubbedInstance, stubInterface } from 'ts-sinon';
+import { StubbedInstance, stubInterface } from 'ts-sinon';
 import { ensureMaster } from '../../../testing/helpers';
 import { MongodSetup, skipIfServerVersion, startTestCluster, skipIfApiStrict } from '../../../testing/integration-testing-hooks';
 import { CliServiceProvider } from '../../service-provider-server';
@@ -24,6 +24,42 @@ chai.use(sinonChai);
 
 function deepClone<T>(value: T): T {
   return JSON.parse(JSON.stringify(value));
+}
+
+
+type ReconfigWithRetryFn = (config: Partial<ReplSetConfig>, options?: Document) => Promise<Document>;
+
+function createRetriableReconfig(rs: ReplicaSet, totalRetries = 12): ReconfigWithRetryFn {
+  // Retriable reconfig function to make sure that test cleanups that deals
+  // with replica set reconfig don't fail because of network issues
+  // Note: We use this helper primarily to aid in test cleanups
+  return async function reconfigWithRetry(config, options = {}) {
+    let result: [ 'error', Error ] | [ 'success', Document ] = [ 'success', {} ];
+    let sleepInterval = 1000;
+    for (let i = 0; i < totalRetries; i++) {
+      try {
+        if (result[0] === 'error') {
+          // Do a mild exponential backoff. If it's been a while since the last
+          // update, also tell the user that we're actually still working on
+          // the reconfig.
+          await rs._instanceState.shellApi.sleep(sleepInterval);
+          sleepInterval *= 1.3;
+          if (sleepInterval > 2500) {
+            await rs._instanceState.shellApi.print('Reconfig did not succeed yet, starting new attempt...');
+          }
+        }
+        result = [ 'success', await rs.reconfig(config, options) ];
+        break;
+      } catch (err: any) {
+        result = [ 'error', err ];
+      }
+    }
+
+    if (result[0] === 'error') {
+      throw result[1];
+    }
+    return result[1];
+  };
 }
 
 describe('ReplicaSet', () => {
@@ -240,70 +276,6 @@ describe('ReplicaSet', () => {
             force: true
           }
         );
-      });
-
-      describe('retry on errors', () => {
-        let oldConfig: Partial<ReplSetConfig>;
-        let reconfigCalls: ReplSetConfig[];
-        let reconfigResults: Document[];
-        let sleepStub: any;
-
-        beforeEach(() => {
-          sleepStub = sinon.stub();
-          instanceState.shellApi.sleep = sleepStub;
-          reconfigCalls = [];
-
-          // eslint-disable-next-line @typescript-eslint/require-await
-          serviceProvider.runCommandWithCheck.callsFake(async(db: string, cmd: Document) => {
-            if (cmd.replSetGetConfig) {
-              return { config: { ...oldConfig, version: oldConfig.version ?? 1 } };
-            }
-            if (cmd.replSetReconfig) {
-              const result = reconfigResults.shift();
-              reconfigCalls.push(deepClone(cmd.replSetReconfig));
-              if (result.ok) {
-                return result;
-              }
-              oldConfig = { ...oldConfig, version: (oldConfig.version ?? 1) + 1 };
-              throw new Error(`Reconfig failed: ${JSON.stringify(result)}`);
-            }
-          });
-        });
-
-        it('does three reconfigs if the first two fail due to known issue', async() => {
-          oldConfig = deepClone(configDoc);
-          reconfigResults = [ { ok: 0 }, { ok: 0 }, { ok: 1 } ];
-
-          const origConfig = deepClone(configDoc);
-          await rs.reconfig(configDoc);
-          expect(reconfigCalls).to.deep.equal([
-            { ...origConfig, version: 2 },
-            { ...origConfig, version: 3 },
-            { ...origConfig, version: 4 }
-          ]);
-          expect(sleepStub).to.have.been.calledWith(1000);
-          expect(sleepStub).to.have.been.calledWith(1300);
-        });
-
-        it('gives up after a number of attempts', async() => {
-          oldConfig = deepClone(configDoc);
-          reconfigResults = [...Array(20).keys()].map(() => ({ ok: 0 }));
-          try {
-            await rs.reconfig(configDoc);
-            expect.fail('missed exception');
-          } catch (err: any) {
-            expect(err.message).to.equal('Reconfig failed: {"ok":0}');
-          }
-          expect(evaluationListener.onPrint).to.have.been.calledWith([
-            await toShellResult('Reconfig did not succeed yet, starting new attempt...')
-          ]);
-          const totalSleepLength = sleepStub.getCalls()
-            .map(({ firstArg }) => firstArg)
-            .reduce((x, y) => x + y, 0);
-          // Expect to spend about a minute sleeping here.
-          expect(totalSleepLength).to.be.closeTo(60_000, 5_000);
-          expect(reconfigCalls).to.have.lengthOf(12);
-        });
       });
     });
     describe('status', () => {
@@ -688,11 +660,8 @@ describe('ReplicaSet', () => {
       let oldConfig: ReplSetConfig;
       let reconfigCalls: ReplSetConfig[];
       let reconfigResults: Document[];
-      let sleepStub: any;
 
       beforeEach(() => {
-        sleepStub = sinon.stub();
-        instanceState.shellApi.sleep = sleepStub;
         secondary = {
           _id: 2, host: 'secondary.mongodb.net', priority: 1, votes: 1
         };
@@ -746,7 +715,7 @@ describe('ReplicaSet', () => {
         }
       });
 
-      it('fails if old note had votes', async() => {
+      it('fails if old node had votes', async() => {
         oldConfig.members.push(secondary);
         try {
           await rs.reconfigForPSASet(2, config);
@@ -795,52 +764,6 @@ describe('ReplicaSet', () => {
           await toShellResult('Running second reconfig to give member at index 2 { priority: 1 }')
         ]);
       });
-
-      it('does three reconfigs the second one fails', async() => {
-        reconfigResults = [{ ok: 1 }, { ok: 0 }, { ok: 1 }];
-        const origConfig = deepClone(config);
-        await rs.reconfigForPSASet(2, config);
-        expect(reconfigCalls).to.deep.equal([
-          { ...origConfig, members: [ config.members[0], config.members[1], { ...secondary, priority: 0 } ], version: 2 },
-          { ...origConfig, version: 3 },
-          { ...origConfig, version: 3 }
-        ]);
-        expect(evaluationListener.onPrint).to.have.been.calledWith([
-          await toShellResult('Running first reconfig to give member at index 2 { votes: 1, priority: 0 }')
-        ]);
-        expect(evaluationListener.onPrint).to.have.been.calledWith([
-          await toShellResult('Running second reconfig to give member at index 2 { priority: 1 }')
-        ]);
-        expect(sleepStub).to.have.been.calledWith(1000);
-      });
-
-      it('gives up after a number of attempts', async() => {
-        reconfigResults = [...Array(20).keys()].map((i) => ({ ok: i === 0 ? 1 : 0 }));
-        try {
-          await rs.reconfigForPSASet(2, config);
-          expect.fail('missed exception');
-        } catch (err: any) {
-          expect(err.message).to.equal('Reconfig failed: {"ok":0}');
-        }
-        expect(evaluationListener.onPrint).to.have.been.calledWith([
-          await toShellResult('Running first reconfig to give member at index 2 { votes: 1, priority: 0 }')
-        ]);
-        expect(evaluationListener.onPrint).to.have.been.calledWith([
-          await toShellResult('Running second reconfig to give member at index 2 { priority: 1 }')
-        ]);
-        expect(evaluationListener.onPrint).to.have.been.calledWith([
-          await toShellResult('Reconfig did not succeed yet, starting new attempt...')
-        ]);
-        expect(evaluationListener.onPrint).to.have.been.calledWith([
-          await toShellResult('Second reconfig did not succeed, giving up')
-        ]);
-        const totalSleepLength = sleepStub.getCalls()
-          .map(({ firstArg }) => firstArg)
-          .reduce((x, y) => x + y, 0);
-        // Expect to spend about a minute sleeping here.
-        expect(totalSleepLength).to.be.closeTo(60_000, 5_000);
-        expect(reconfigCalls).to.have.lengthOf(1 + 12);
-      });
     });
   });
 
@@ -860,6 +783,7 @@ describe('ReplicaSet', () => {
     let instanceState: ShellInstanceState;
     let db: Database;
     let rs: ReplicaSet;
+    let reconfigWithRetry: ReconfigWithRetryFn;
 
     before(async function() {
       this.timeout(100_000);
@@ -877,6 +801,7 @@ describe('ReplicaSet', () => {
       instanceState = new ShellInstanceState(serviceProvider);
       db = instanceState.currentDb;
       rs = new ReplicaSet(db);
+      reconfigWithRetry = createRetriableReconfig(rs);
 
       // check replset uninitialized
       try {
@@ -967,7 +892,7 @@ describe('ReplicaSet', () => {
         expect(status.version).to.be.greaterThan(version);
       });
       afterEach(async() => {
-        await rs.reconfig(cfg);
+        await reconfigWithRetry(cfg);
         const status = await rs.conf();
         expect(status.members.length).to.equal(3);
       });
@@ -999,7 +924,7 @@ describe('ReplicaSet', () => {
         expect(conf.version).to.be.greaterThan(version);
       });
       afterEach(async() => {
-        await rs.reconfig(cfg);
+        await reconfigWithRetry(cfg);
         const status = await rs.conf();
         expect(status.members.length).to.equal(3);
       });
@@ -1015,7 +940,7 @@ describe('ReplicaSet', () => {
         expect(conf.version).to.be.greaterThan(version);
       });
       afterEach(async() => {
-        await rs.reconfig(cfg);
+        await reconfigWithRetry(cfg);
         const status = await rs.conf();
         expect(status.members.length).to.equal(3);
       });
